move average config to API
