+++ src/main/java/com/xceptance/xlt/api/report/MovingAverageConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.xceptance.xlt.report.util;
+package com.xceptance.xlt.api.report;
 
 import java.util.Arrays;
 import java.util.List;
