+++ src/test/java/com/xceptance/xlt/report/util/JFreeChartUtilsTest.java
@@ -25,6 +25,7 @@ import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import com.xceptance.xlt.api.report.MovingAverageConfiguration;
 import com.xceptance.xlt.report.ReportGeneratorMain;
 
 import junitparams.JUnitParamsRunner;
